Section in README for updating existing installs
I've added a section into the
README that details the process
of updating a previous installation
of bdvl on a target box. This was
already possible previously, but
I had not put it to the test properly
until now.

auto.sh is now a little bit smaller.
Other very tiny changes.