New FILE_STEAL method & tidying up code
There is now an option in bdvl
for prioritising sending stolen
files somewhere offsite. I have
created etc/hoarder.c which
handles files received & writes
files from each connected box,
into their own directory.
etc/hoarder.c has a few options
as to how it runs... daemonizing
it isn't something it does itself,
but might be in the next commit.

In setup.py there is STOLEN_STORE.
See this for comments. There is
also a part in the README which
mentions what this is for.

If NO_DISK_WRITE in setup.py is
True, and if the file contents
cannot be sent for any reason,
the file will not be copied into
the dedicated hidden directory.

If you specify, in setup.py,
somewhere to try to send files
to, you will be able to change
it later on while in the kit's
backdoor home directory. It is
available as a symlink.

bdvl's bdvupdate function has had
a touchup as of the last commit.

Other minor changes too...