Read sopath at runtime, other code updates
The rootkit is now capable of
reading the path of the rootkit
from /proc/$$/maps. This is
always the default way of
determining where the rootkit is,
other than at installation/update...

The bdvupdate function is now
capable of installing without
needing to manually know if
PATCH_DYNAMIC_LINKER is defined
in the target installation.
Now updating is just a matter
of pressing enter a few times.

Other fairly minor changes to
other existing code.. just
tidying things up a bit.